# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "",
    "version": "v0.01",
    "title": "API for online commercial",
    "contact": {
      "name": "Mikhail",
      "email": "example@mail.com"
    }
  },
  "servers": [
    {
      "url": "/rest"
    }
  ],
...
  "paths": {
    "/list": {
      "post": {
        "tags": ["main"],
        "summary": "Return objects",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "description": "query",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryBody"
              }
            }
          }
        }
      }
    },
    "/get": {
      "post": {
        "tags": [
          "main"
        ],
        "summary": "Return object by id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "description": "query",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryBodyForGet"
              }
            }
          }
        }
      }
    },
    "/update": {
      "post": {
        "tags": [
          "main"
        ],
        "summary": "Insert or update object",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "description": "query",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryBodyForUpdate"
              }
            }
          }
        }
      }
  },
    "/delete": {
      "post": {
        "tags": ["main"],
        "summary": "Remove object by id",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "description": "query",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryBodyForGet"
              }
            }
          }
        }
      }
    },
    "/get_size/{table_name}": {
      "get": {
        "tags": ["main"],
        "summary": "Get size",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "table_name",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/categories/get_by_name/{input_name}": {
      "get": {
        "tags": ["main"],
        "summary": "Search in \"categories\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input_name",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/categories/search_by_name/{input_name}": {
      "get": {
        "tags": ["main"],
        "summary": "Partial search in \"categories\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input_name",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/vendors/get_by_name/{input_name}": {
      "get": {
        "tags": ["main"],
        "summary": "Search in \"vendors\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input_name",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/vendors/search_by_name/{input_name}": {
      "get": {
        "tags": ["main"],
        "summary": "Partial search in \"vendors\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input_name",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/products/search_by_category_id/{input}": {
      "get": {
        "tags": ["main"],
        "summary": "Search by \"category_id\" in \"products\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/products/search_by_vendor_id/{input}": {
      "get": {
        "tags": ["main"],
        "summary": "Search by \"vendor_id\" in \"products\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/products/search_by_model/{input}": {
      "get": {
        "tags": ["main"],
        "summary": "Search by \"model\" in \"products\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/products/search_by_specs/{input}": {
      "get": {
        "tags": ["main"],
        "summary": "Search by \"specs\" in \"products\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/stores/get_by_email/{input_email}": {
      "get": {
        "tags": ["main"],
        "summary": "Search by \"email\" in \"stores\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input_email",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/stores/search_by_email/{input_email}": {
      "get": {
        "tags": ["main"],
        "summary": "Partial search by \"email\" in \"stores\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input_email",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
...
    "/clients/search_by_firstname/{input}": {
      "get": {
        "tags": ["main"],
        "summary": "Partial search by \"firstname\" in \"clients\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/clients/search_by_lastname/{input}": {
      "get": {
        "tags": ["main"],
        "summary": "Partial search by \"lastname\" in \"clients\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/clients/search_by_email/{input}": {
      "get": {
        "tags": ["main"],
        "summary": "Partial search by \"email\" in \"clients\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input_paid_delivery",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/clients/search_by_contact_number/{input}": {
      "get": {
        "tags": ["main"],
        "summary": "Search by \"contact_number\" in \"clients\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/orders/search_by_client_id/{input}": {
      "get": {
        "tags": ["main"],
        "summary": "Search by \"client_id\" in \"orders\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "integer"
          }
        }
      ]
      }
    },
    "/orders/search_by_order_id/{input}": {
      "get": {
        "tags": ["main"],
        "summary": "Search by \"store_id\" in \"orders\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "integer"
          }
        }
      ]
      }
    },
    "/orders/search_by_date/{date}": {
      "get": {
        "tags": ["main"],
        "summary": "Search by \"date\" in \"orders\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "date",
          "in": "path",
          "required": true,
          "description": "Format: (YYYY-MM-DD)",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/orders/search_by_time/{left_value}/{right_value}": {
      "get": {
        "tags": ["main"],
        "summary": "Search by \"time\" in \"orders\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "left_value",
          "in": "path",
          "required": true,
          "description": "Format: (HH:MM:SS)",
          "schema": {
            "type": "string"
          }
        },
          {
          "name": "right_value",
          "in": "path",
          "required": true,
          "description": "Format: (HH:MM:SS)",
          "schema": {
            "type": "string"
          }
        }

      ]
      }
    },
    "/orders/search_by_confirmation/{input_confirmation}": {
      "get": {
        "tags": ["main"],
        "summary": "Search by \"confirmation\" in \"orders\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "boolean"
          }
        }
      ]
      }
    },
    "/couriers/search_by_firstname/{input}": {
      "get": {
        "tags": ["main"],
        "summary": "Partial search by \"firstname\" in \"couriers\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/couriers/search_by_lastname/{input}": {
      "get": {
        "tags": ["main"],
        "summary": "Partial search by \"lastname\" in \"couriers\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/couriers/search_by_email/{input}": {
      "get": {
        "tags": ["main"],
        "summary": "Partial search by \"email\" in \"couriers\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input_paid_delivery",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
    "/couriers/search_by_contact_number/{input}": {
      "get": {
        "tags": ["main"],
        "summary": "Search by \"contact_number\" in \"couriers\"",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
        {
          "name": "input",
          "in": "path",
          "required": true,
          "description": "",
          "schema": {
            "type": "string"
          }
        }
      ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "QueryBody": {
        "type": "object",
        "properties": {
          "table_name": {
            "type": "string",
            "description": "table name",
            "default": "vendors"
          },
          "offset_value": {
            "type": "integer",
            "description": "fetch value",
            "default": 0
          },
          "limit_value": {
            "type": "integer",
            "description": "limit value",
            "default": 10
          }
        }
      },
      "QueryBodyForGet": {
        "type": "object",
        "properties": {
          "table_name": {
            "type": "string",
            "description": "table name",
            "default": "vendors"
          },
          "id": {
            "type": "integer",
            "description": "id",
            "default": "1"
          }
        }
      },
      "QueryBodyForUpdate": {
        "type": "object",
        "properties": {
          "table_name": {
            "type": "string",
            "description": "table name",
            "default": "vendors"
          },
          "obj": {
            "type": "object",
            "$ref": "#/components/schemas/VendorBody"
          }
        }
      },
...
      "VendorBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "if it new object must be -1",
            "default": -1
          },
          "name": {
            "type": "string",
            "description": "not empty"
          }
        }
      },
...
    }
  }
}